+++ controllers/operator/construct/database_construction.go
@@ -812,7 +812,7 @@ func defaultPodAnnotations(certHash string) map[string]string {
 		// this annotation is necessary in order to trigger a pod restart
 		// if the certificate secret is out of date. This happens if
 		// existing certificates have been replaced/rotated/renewed.
-		"certHash": certHash,
+		certs.CertHashAnnotationKey: certHash,
 	}
 }